+++ pydantic_ai_slim/pydantic_ai/providers/cerebras.py
@@ -8,10 +8,7 @@ import httpx
 from pydantic_ai import ModelProfile
 from pydantic_ai.exceptions import UserError
 from pydantic_ai.models import cached_async_http_client
-from pydantic_ai.profiles.harmony import harmony_model_profile
-from pydantic_ai.profiles.meta import meta_model_profile
-from pydantic_ai.profiles.openai import OpenAIJsonSchemaTransformer, OpenAIModelProfile
-from pydantic_ai.profiles.qwen import qwen_model_profile
+from pydantic_ai.profiles.cerebras import cerebras_provider_model_profile
 from pydantic_ai.providers import Provider
 
 try:
@@ -21,35 +18,6 @@ except ImportError as _import_error:  # pragma: no cover
         'Please install the `openai` package to use the Cerebras provider, '
         'you can use the `openai` optional group — `pip install "pydantic-ai-slim[openai]"`'
     ) from _import_error
-
-
-def cerebras_provider_model_profile(model_name: str) -> ModelProfile | None:
-    """Get the model profile for a model routed through Cerebras provider.
-
-    This function handles model profiling for models that use Cerebras's API,
-    and applies Cerebras-specific settings like unsupported model parameters.
-    """
-    prefix_to_profile = {'llama': meta_model_profile, 'qwen': qwen_model_profile, 'gpt-oss': harmony_model_profile}
-
-    profile = None
-    for prefix, profile_func in prefix_to_profile.items():
-        model_name = model_name.lower()
-        if model_name.startswith(prefix):
-            profile = profile_func(model_name)
-
-    # According to https://inference-docs.cerebras.ai/resources/openai#currently-unsupported-openai-features,
-    # Cerebras doesn't support some model settings.
-    unsupported_model_settings = (
-        'frequency_penalty',
-        'logit_bias',
-        'presence_penalty',
-        'parallel_tool_calls',
-        'service_tier',
-    )
-    return OpenAIModelProfile(
-        json_schema_transformer=OpenAIJsonSchemaTransformer,
-        openai_unsupported_model_settings=unsupported_model_settings,
-    ).update(profile)
 
 
 class CerebrasProvider(Provider[AsyncOpenAI]):
