+++ pydantic_ai_slim/pydantic_ai/providers/groq.py
@@ -8,14 +8,7 @@ import httpx
 from pydantic_ai import ModelProfile
 from pydantic_ai.exceptions import UserError
 from pydantic_ai.models import cached_async_http_client
-from pydantic_ai.profiles.deepseek import deepseek_model_profile
-from pydantic_ai.profiles.google import google_model_profile
-from pydantic_ai.profiles.groq import groq_model_profile
-from pydantic_ai.profiles.meta import meta_model_profile
-from pydantic_ai.profiles.mistral import mistral_model_profile
-from pydantic_ai.profiles.moonshotai import moonshotai_model_profile
-from pydantic_ai.profiles.openai import openai_model_profile
-from pydantic_ai.profiles.qwen import qwen_model_profile
+from pydantic_ai.profiles.groq import groq_provider_model_profile
 from pydantic_ai.providers import Provider
 
 try:
@@ -25,51 +18,6 @@ except ImportError as _import_error:  # pragma: no cover
         'Please install the `groq` package to use the Groq provider, '
         'you can use the `groq` optional group — `pip install "pydantic-ai-slim[groq]"`'
     ) from _import_error
-
-
-def groq_moonshotai_model_profile(model_name: str) -> ModelProfile | None:
-    """Get the model profile for an MoonshotAI model used with the Groq provider."""
-    return ModelProfile(supports_json_object_output=True, supports_json_schema_output=True).update(
-        moonshotai_model_profile(model_name)
-    )
-
-
-def meta_groq_model_profile(model_name: str) -> ModelProfile | None:
-    """Get the model profile for a Meta model used with the Groq provider."""
-    if model_name in {'llama-4-maverick-17b-128e-instruct', 'llama-4-scout-17b-16e-instruct'}:
-        return ModelProfile(supports_json_object_output=True, supports_json_schema_output=True).update(
-            meta_model_profile(model_name)
-        )
-    else:
-        return meta_model_profile(model_name)
-
-
-def groq_provider_model_profile(model_name: str) -> ModelProfile | None:
-    """Get the model profile for a model routed through Groq provider.
-
-    This function handles model profiling for models that use Groq's API,
-    including various model families like Llama, Gemma, Qwen, etc.
-    """
-    prefix_to_profile = {
-        'llama': meta_model_profile,
-        'meta-llama/': meta_groq_model_profile,
-        'gemma': google_model_profile,
-        'qwen': qwen_model_profile,
-        'deepseek': deepseek_model_profile,
-        'mistral': mistral_model_profile,
-        'moonshotai/': groq_moonshotai_model_profile,
-        'compound-': groq_model_profile,
-        'openai/': openai_model_profile,
-    }
-
-    for prefix, profile_func in prefix_to_profile.items():
-        model_name = model_name.lower()
-        if model_name.startswith(prefix):
-            if prefix.endswith('/'):
-                model_name = model_name[len(prefix) :]
-            return profile_func(model_name)
-
-    return None
 
 
 class GroqProvider(Provider[AsyncGroq]):
